properties: support recursive expansion with ${+...}
Any normal ${...} section may now also have one '+' after the left brace,
which denotes to re-expand the result string of that section.

E.g. it's now possible to use --term-status-msg='${+title}' which would
first expand the title variable to the template which it is, and then expand
it again to actually apply the template. Effectively this would make the
term status message identical to the window title, including after the title
is set to be a different template. This was not possible without re-expand.

The re-expand may also expand further ${+...} etc. The recursion depth is
limited arbitrarily to 10 - only for a variable inside a ${+...} section which
after expansion includes another such. Every valid prior expression is still
valid, specifically nested expansion (STR) are not limited (except by the
string length - like before).